api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Daniel Ale <danielale9291@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.0"
djangorestframework = "^3.14.0"
markupsafe = "^2.1.3"
markdown = "^3.5.1"
django-import-export = "^3.3.5"
flower = "^2.0.1"
django-redis = "^5.4.0"
redis = "^5.0.1"
celery = "^5.3.6"
channels-redis = "^4.1.0"
channels = "^4.0.0"
sendgrid = "^6.11.0"
pillow = "^10.1.0"
packaging = "^23.2"
django-storages = "^1.14.2"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
botocore = "^1.34.9"
pyotp = "^2.9.0"
cuid2 = "^2.0.0"
gunicorn = "^21.2.0"
daphne = "^4.0.0"
psycopg2-binary = "^2.9.9"
django-jazzmin = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
